+++ lib/elliptic.dart
@@ -8,3 +8,4 @@ export 'src/base.dart';
 export 'src/publickey.dart';
 export 'src/privatekey.dart';
 export 'src/elliptic.dart';
+export 'src/curves.dart';