+++ test/ecdsa_test.dart
@@ -1,4 +1,5 @@
 import 'package:elliptic/elliptic.dart';
+import 'package:elliptic/src/curves.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -157,6 +158,38 @@ void main() {
           PublicKey.fromHex(ec,
               '048fe605cc744b71bfba25bb69acc4c29a1a4a8c758d8262dc9bcf3472ce0f50c532e2e4461fc1b8cc61b5f705962aad28300246592e507d2778d3ca2eac78e1fb'),
           equals(priv3.publicKey));
+    });
+
+    test('Test S-256', () {
+      var testVectors = [
+        [
+          'c37c299bb7f5ffd8d9329d052983342a8c3234ff3b3fa32a292187341f7146d7',
+          '03a12b6218425127f186011ff4c203b8d6ea651877c46f12484b2eda492596484f'
+        ],
+        [
+          '52d62cfcf7062af53f7bec124fe9285eaa8a8963411ba613b7432be73565b6b3',
+          '0449aedf74e8f87811761cf3d5fa8f8eaa42b4c657efc986939229a898b3ee27a000f9a247290fb5716c2db17dc193fd4e3f7b36a9947b477c6f0769c15fb8bf79'
+        ],
+        [
+          '0df8d8c6a2bb48af1be87d8403afc15917962898a4fad2b2584276f6adcf7ced',
+          '044436b31b9d663d2bb7d6d6e56678f81ed0db65ee74c123ef238b1678399f77e42985555da48e7a80198a5aec11c369ce7b63a759a492ac4315dbf083e6668ce6'
+        ],
+        [
+          '8db69356f772b318c523bbcaa5bddfeddae118ee6aca574e7d2e4332e35fc238',
+          '0247d3faa09ce8c4bd46c9b89a680d5b1064bf2bcd321a47358c02e9527c2387e6'
+        ]
+      ];
+
+      var curve = getSecp256k1();
+      for (var vec in testVectors) {
+        var priv = PrivateKey.fromHex(curve, vec[0]);
+        var pub1 = priv.publicKey;
+        var pub2 = PublicKey.fromHex(curve, vec[1]);
+        print(pub1);
+        print(pub2);
+        expect(curve.isOnCurve(pub1), true);
+        expect(pub1, pub2);
+      }
     });
   });
 }